Register TearDown* methods after Setup* methods
This allows the TearDown* methods to run before any cleanup methods
registered within the Setup* methods.

For example, when using pgtestdb you can create a test-specific DB
inside SetupTest which will register a cleanup function to drop the
DB at the end of the test, but the DB cannot be dropped until the
connection is closed in TearDownTest.